added namespaces and types to base models
