+++ library/models/BaseRequestModel.php
@@ -20,9 +20,12 @@
  * Mother of all requests. For now with every request, the  credentials have to be sent as well.
  * This will be changed to oAuth in the near future.
  */
+ 
+namespace Barion\Models;
+ 
 class BaseRequestModel
 {
     // for poskey authentication
-    public $POSKey;
+    public string $POSKey;
 
 }
